src/modules/event_hub/event_hub.service.spec.ts
import { Test, TestingModule } from '@nestjs/testing';
import { ServiceBusService } from '../service_bus/service_bus.service';
import { EventHubService } from './event_hub.service';

// Mock dependencies
jest.mock('@azure/event-hubs');
jest.mock('@azure/storage-blob');
jest.mock('../service_bus/service_bus.service');

describe('EventHubService', () => {
  let service: EventHubService;
  let serviceBusService: ServiceBusService;

  beforeEach(async () => {
    const module: TestingModule = await Test.createTestingModule({
      providers: [EventHubService, ServiceBusService],
    }).compile();

    service = module.get<EventHubService>(EventHubService);
    serviceBusService = module.get<ServiceBusService>(ServiceBusService);
  });

  it('should be defined', () => {
    expect(service).toBeDefined();
  });

  describe('SendToEventHub', () => {
    it('should send events to the Event Hub', async () => {
      // Mock environment variables
      process.env.EVENT_HUB_CONNECTION_STRING = 'mockConnectionString';
      process.env.EVENT_HUB_NAME = 'mockEventHubName';

      // Execute the method
      await service.SendToEventHub();

      // Verify that events are sent
      expect(serviceBusService.sendToServiceBusQueue).toHaveBeenCalledTimes(3);
    });
  });

  describe('ReceiveFromEventHub', () => {
    it('should receive events from the Event Hub', async () => {
      // Mock environment variables
      process.env.EVENT_HUB_CONNECTION_STRING = 'mockConnectionString';
      process.env.EVENT_HUB_NAME = 'mockEventHubName';
      process.env.EVENT_HUB_CONSUMER_GROUP = 'mockConsumerGroup';
      process.env.CONTAINER_CONNECTION_STRING = 'mockStorageConnectionString';
      process.env.CONTAINER_NAME = 'mockContainerName';

      // Execute the method
      await service.ReceiveFromEventHub();

      // Assert that the subscription is set up
      // Note: This test might need more specific assertions based on actual behavior.
      expect(true).toBe(true);
    });
  });
});
